cursor management corrected
